Create Sokoban
Task Description: The program's input is a description of the board's initial state and a non-empty sequence of commands, terminated by a dot. The wireless output signal represents the board's state.

The program skips input after the dot that terminates the data.

The board description consists of non-empty lines containing single-character representations of the shelf states:

-

an empty square that is not a deletion square;

+

an empty square that is a deleted square;

#

a wall;

@

an empty square, not a message notification, on which the character is located;

*

an empty target pole on which the character is located;

[a..z] (lowercase letter)

a square that is not a message square, on which the chest with the given name is located;

[A..Z] (uppercase letter)

a target square on which the chest with the given name is located, corresponding to the given uppercase letter.

There is exactly one character in a valid board description. The lower chest, marked with the appropriate letter of the alphabet, can only appear once.